Added the error messages required and fixed problem the Authcontext
Because with Cypress I couldn't clear the sessionStorage, so when I rerun all the test,
the application was broken. The reason of this, whenever I run a single test I clean the localStorage,
But the session of the test is not changing, so the sessionStorage is persistent within the test but I
need to clear the session or change the way of making the code.